pyproject.toml
[tool.poetry]
name = "aiogram-book"
version = "0.1.0"
description = "Telegram chat-bot based on aiogram for online reading book"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.1"
dishka = "^1.3.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
python-decouple = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
